manifests: Serve oc from within the cluster
We need to support the ability to download clients (oc) from the
cluster when an install is successful.  With this commit, I'm adding a
new Deployment for serving the binary from the cli image referenced
from the release image.  That way, the provided oc is the appropriate
version for the current cluster version.  I've set 'replicas' to two
to support continued availability during rolling updates.

The Python here-document approach is based on
openshift/ci-operator@95a41a39 (Use a multithreaded python server to
serve RPMs, 2018-06-25, openshift/ci-operator#35).  The script creates
a scratch directory to serve from, and populates it with the oc binary
(via a symlink).  It also creates zip and tar versions, so users can
do things like:

  $ wget https://downloads-openshift-console.apps.wking.devcluster.openshift.com/amd64/linux/oc.zip
  $ unzip oc.zip

instead of:

  $ wget https://downloads-openshift-console.apps.wking.devcluster.openshift.com/amd64/linux/oc
  $ chmod +x oc

if they prefer.  We don't have to clean up the scratch directory in
the script, because it will be reaped as part of container cleanup.

Moving forward, we'll want to replace the cli image with a
cli-downloads image that includes builds for other platforms and
architectures, but I'm punting on that for now.

[1]: https://github.com/openshift/ci-operator/commit/95a41a395fa778680b3dc93c845407eb8a1a0ae1