Improve cluster operator status reporting during upgrades
We should show the version as-is instead of stripping everything after
the `-`. Note for release versions, the version number will be short
like `4.1.1`. For prerelease builds, stripping content after `-` removes
the unique part of the version number.

We should also show informative messages in the cluster operator status
even when there are no errors.