Enable multi-arch support in downloads
In order for console-operator deployment to succeed, cli-artifacts needs
to be available on all arches for downloads-openshift-console. However,
in that case, /usr/bin/oc (inherited from cli) is a native binary, and
we want to provide all primary Linux architectures to match those on
mirror.openshift.com, regardless of cluster architecture.

This depends on https://github.com/openshift/oc/pull/153 to provide the
new binaries in cli-artifacts.