manifests/07-downloads-deployment: Serve from standardized paths
Lean on the path consistency from openshift/oc@e949088d (Enable all
Linux arches in cli-artifacts, 2019-11-07, openshift/oc#153) to make
it easier to scale as we add/remove arch/OS support.  With this
commit, the downloads deployment is less opinionated about what is in
the cli-artifacts bundle.  It looks for all possible arch/OS
combinations, serves what it finds, and 404s requests for binaries it
could not find.

I've also renamed mac -> darwin in the canonical URIs, so there's no
mental overhead for translating from the conventional GOOS/GOARCH [1].
I've added symlinks in the pod to support external folks who had
hard-coded the old URIs, but we should be able to drop those symlinks
after a reasonable deprecation period.  I have not overhauled the
server to 301 requests to the deprecated URI, and we probably want
that before dropping the symlinks.

[1]: https://golang.org/doc/install/source#environment