pkg/console/controllers/clidownloads/controller: Arch qualifiers for Mac/Windows
Folks could presumably be running these OSes on other architectures.
For example, the GOOS/GOARCH docs [1] provide for darwin on 386,
amd64, arm, and arm64 and windows on 386 and amd64 (although Windows
runs on ARM64 [2] and possibly other architectures as well).

I'm sticking with x86_64 for consistency with the Linux entry from
865d84d7bb (Enable multi-arch Linux links in clidownloads, 2019-11-08, #343),
although Linus says the x86-64 form predates the underscore form [3].

[1]: https://golang.org/doc/install/source#environment
[2]: https://docs.microsoft.com/en-us/windows/arm/
[3]: https://lkml.org/lkml/2004/2/21/110