Pause MachineConfigPools during proxy test
The MCO's behavior changed such that it is actually now capable of
reacting to and rolling out the fake proxy configuration supplied by the
proxy test, which causes the kubelet to fail on the way back up after
the nodes are rebooted.
See: https://issues.redhat.com/browse/OCPBUGS-5780

This uses the MCO's pool pause feature to pause all machine config pools
while the proxy test is taking place, essentially making this test
behave as it previously (before the MCO change). It does add a few
seconds to the run time of TestProxy.