Actually wait for the machineconfig
This test can happen so fast that we pause and unpause before the config
even gets rendered, which means it still makes it out to the nodes and
degrades them after this test is over.

This now explicitly makes us wait for those configurations to render, so
we know where we are in the process and nothing can sneak past us.