Dockerfile: Be less opinionated about CRD filesnames when copying from openshift/api
openshift/api@9e3e820a70e4b8 (Console CRDs: Rename files to use the
default run-level for updates, 2023-08-02, openshift/api#1541) renamed
these, and needing opinions in two locations that must match makes
maintenance tedious.  With this commit, I'm asserting that
openshift/api is where CRD filename opinions happen, and that this
repo just passes those filenames along.  And I'm asserting that the
CRDs we're responsible for are "all of console/ and helm/" and "any
CRDs in operator/ that include 'console-operator' in their filename".