RBAC: add "watch secrets in openshift-config-managed"
The secret syncer is set up for CMs and Secrets from openshift-config
and openshift-config-managed. However, it only syncs CMs from the
openshift-config-managed.

We may not want to sync secrets from there so we could use the previous
hack removing the secrets client, and create an additional secrets
syncer from only `openshift-config` instead.