The console-deployment should set the number of replicas based on the ControlPlaneTopology
The node selector for the console deployment requires deploying it on the master nodes, The node selector for the console deployment requires deploying it on the master nodes, while the replica count is determined by the infrastructureTopology, which primarily tracks the workers' setup.

When an OpenShift cluster is installed with a single master node and multiple workers, this leads the console deployment to request 3 replicas as infrastructureTopology is set to HighlyAvailable. Instead, ControlPlaneTopology is set to SingleReplica as expected.

This commit makes the operator drive the number of replicas for the console deployment based on the ControlPlaneTopology rather than the InfrastructureTopology.