(bugfix): don't set current clients when no OIDC providers are configured
by resetting the AuthStatusHandler.currentClientID to an empty
string when the Authentication.Spec.Type is not set to OIDC.
Additionally, protect against a degraded state when
AuthStatusHandler.currentClientID is set but there is no
evidence of OIDC providers being configured in the spec by
only adding current clients when both the currentClientID
is set and there is evidence of OIDC providers being configured.
Add a warning log when we notice the currentClientID is set but
there is no evidence of OIDC providers being configured.

Fixes [OCPBUGS-44953](https://issues.redhat.com/browse/OCPBUGS-44953)

Signed-off-by: everettraven <everettraven@gmail.com>